+++ Tests/RunnerLibTests/DangerJSVersionFinderTests.swift
@@ -4,13 +4,13 @@ import XCTest
 final class DangerJSVersionFinderTests: XCTestCase {
     func testItSendsTheCorrectCommandAndReturnsTheCorrectResult() throws {
         let executor = MockedExecutor()
-        executor.result = "1.0.0"
+        executor.result = { _ in "1.0.0" }
 
         let dangerJSPath = "/test/danger"
 
         let version = DangerJSVersionFinder.findDangerJSVersion(dangerJSPath: dangerJSPath, executor: executor)
 
-        XCTAssertEqual(executor.receivedCommand, dangerJSPath + " --version")
-        XCTAssertEqual(version, executor.result)
+        XCTAssertEqual(executor.receivedCommands, [dangerJSPath + " --version"])
+        XCTAssertEqual(version, "1.0.0")
     }
 }