+++ Tests/RunnerLibTests/GetDangerJSPathTests.swift
@@ -7,19 +7,34 @@ final class GetDangerJSPathTests: XCTestCase {
         Logger(isVerbose: false, isSilent: false, printer: SpyPrinter())
     }
 
-    func testItUsesTheDangerJSPathOptionIfPresent() throws {
+    func testItUsesDangerJSPathOptionIfPresent() throws {
         let expectedResult = "/franco/test/danger-js"
         let path = try getDangerCommandPath(logger: logger, args: ["test", "--danger-js-path", expectedResult])
         XCTAssertEqual(path, expectedResult)
     }
 
-    func testItSearchesForDangerJSIfTheDangerJSPathOptionIsNotPresent() throws {
+    func testItSearchesForDangerJSIfDangerJSPathOptionIsNotPresent() throws {
         let executor = MockedExecutor()
-        let expectedResult = "/usr/test/danger-js"
-        executor.result = expectedResult
+        executor.result = { _ in "/usr/test/danger-js" }
 
         let path = try getDangerCommandPath(logger: logger, args: [], shellOutExecutor: executor)
-        XCTAssertEqual(executor.receivedCommand, "command -v danger")
+        XCTAssertEqual(executor.receivedCommands, ["command -v danger-js"])
+        XCTAssertEqual(path, "/usr/test/danger")
+    }
+
+    func testItSearchesForDangerIfTheDangerPathOptionIsNotPresentAndDangerJSIsNotFound() throws {
+        let executor = MockedExecutor()
+        let expectedResult = "/usr/test/danger"
+        executor.result = { command in
+            if command.hasSuffix("danger-js") {
+                return ""
+            } else {
+                return expectedResult
+            }
+        }
+
+        let path = try getDangerCommandPath(logger: logger, args: [], shellOutExecutor: executor)
+        XCTAssertEqual(executor.receivedCommands, ["command -v danger-js", "command -v danger"])
         XCTAssertEqual(path, expectedResult)
     }
 }