+++ Package.swift
@@ -9,7 +9,7 @@ let package = Package(
     products: [
         .library(name: "Danger", type: .dynamic, targets: ["Danger"]),
         .library(name: "DangerFixtures", type: .dynamic, targets: ["DangerFixtures"]),
-        .library(name: "DangerDeps", type: .dynamic, targets: ["Danger-Swift"]), // dev
+//        .library(name: "DangerDeps", type: .dynamic, targets: ["Danger-Swift"]), // dev
         .executable(name: "danger-swift", targets: ["Runner"]),
     ],
     dependencies: [
@@ -18,24 +18,24 @@ let package = Package(
         .package(url: "https://github.com/nerdishbynature/octokit.swift", from: "0.10.1"),
         // Danger Plugins
         // Dev dependencies
-        .package(url: "https://github.com/shibapm/Komondor", from: "1.0.0"), // dev
-        .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.35.8"), // dev
-        .package(url: "https://github.com/Realm/SwiftLint", from: "0.38.0"), // dev
-        .package(url: "https://github.com/pointfreeco/swift-snapshot-testing.git", from: "1.7.1"), // dev
-        .package(url: "https://github.com/shibapm/Rocket", from: "0.4.0"), // dev
-        .package(url: "https://github.com/SwiftDocOrg/swift-doc", .branch("master")), // dev
+//        .package(url: "https://github.com/shibapm/Komondor", from: "1.0.0"), // dev
+//        .package(url: "https://github.com/nicklockwood/SwiftFormat", from: "0.35.8"), // dev
+//        .package(url: "https://github.com/Realm/SwiftLint", from: "0.38.0"), // dev
+//        .package(url: "https://github.com/pointfreeco/swift-snapshot-testing.git", from: "1.7.1"), // dev
+//        .package(url: "https://github.com/shibapm/Rocket", from: "0.4.0"), // dev
+//        .package(url: "https://github.com/SwiftDocOrg/swift-doc", .branch("master")), // dev
     ],
     targets: [
-        .target(name: "Danger-Swift", dependencies: ["Danger"]), // dev
+//        .target(name: "Danger-Swift", dependencies: ["Danger"]), // dev
         .target(name: "DangerShellExecutor"),
         .target(name: "DangerDependenciesResolver", dependencies: ["DangerShellExecutor", "Version", "Logger"]),
         .target(name: "Danger", dependencies: ["OctoKit", "Logger", "DangerShellExecutor"]),
         .target(name: "RunnerLib", dependencies: ["Logger", "DangerShellExecutor"]),
         .target(name: "Runner", dependencies: ["RunnerLib", "Logger", "DangerDependenciesResolver"]),
         .target(name: "DangerFixtures", dependencies: ["Danger"]),
-        .testTarget(name: "DangerTests", dependencies: ["Danger", "DangerFixtures", "SnapshotTesting"]), // dev
-        .testTarget(name: "RunnerLibTests", dependencies: ["RunnerLib", "SnapshotTesting"]), // dev
-        .testTarget(name: "DangerDependenciesResolverTests", dependencies: ["DangerDependenciesResolver", "SnapshotTesting"]), // dev
+//        .testTarget(name: "DangerTests", dependencies: ["Danger", "DangerFixtures", "SnapshotTesting"]), // dev
+//        .testTarget(name: "RunnerLibTests", dependencies: ["RunnerLib", "SnapshotTesting"]), // dev
+//        .testTarget(name: "DangerDependenciesResolverTests", dependencies: ["DangerDependenciesResolver", "SnapshotTesting"]), // dev
     ]
 )